pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eduacdc"
version = "0.1.0"
description = "eduACDC - Educational Python implementation of ACDC (Atmospheric Cluster Dynamics Code)"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "GPL-3.0"}

dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "sympy>=1.9.0",
    "matplotlib>=3.5.0",
    "pyyaml>=6.0",
    "tqdm>=4.62.0",
    "pandas>=1.3.0",
    "pint>=0.20.0",
    "networkx>=3.1",
    "seaborn>=0.13.2",
]

[tool.setuptools.packages.find]
include = ["eduacdc*"]

[project.optional-dependencies]
dev = [
    "jupyter>=1.1.1",
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "pytest-mock>=3.6",
    "pytest-xdist>=2.0",
    "pre-commit>=2.15",
]
